% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Best.Index.R, R/Jack.Indices.R
\name{Best.Index}
\alias{Best.Index}
\title{Jack-knife indices in a single topology m times and evaluates a success rule.}
\usage{
Best.Index(tree = tree, distrib = distrib, jtip = jtip,
  replicates = replicates, success = c(success))

Best.Index(tree = tree, distrib = distrib, jtip = jtip,
  replicates = replicates, success = c(success))
}
\description{
The function jack-knifes the terminals jtip (=replicates) times, and calculates the number of times 
that each replicate recovers the same initial initial.Rankinging (X:Y positions), by default success is 1:3 positions.
The function returns the number of hits.

The function jack-knifes the terminals and
calculates the indices value m (=replicates) times.  
The function returns the success that correspond to  
obtain the same ranking for X,Y positions, established as the vector
success (by default success)).
}

