% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Calculate.Index}
\alias{Calculate.Index}
\title{Indices values for a single topology}
\usage{
Calculate.Index(tree = tree, distrib = distrib)
}
\description{
The funtion calculates standard indices I and W,
along with Posadas et al. (2001) modifications.
}

