% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{distrib.figure1}
\alias{distrib.figure1}
\title{Figure 1 distribution}
\format{A data.framer object with eight areas (columns) and five species (rows)}
\usage{
distrib.figure1
}
\description{
A data frame with five species distributions in eight areas.
}
\keyword{datasets}

