% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{distribution}
\alias{distribution}
\title{Figure 1 distribution}
\format{A data.frame object with eight areas (columns) and five species (rows)}
\usage{
distribution
}
\description{
A data frame with five species distributions in eight areas.
}
\keyword{datasets}

