% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexW.R
\name{IndexW}
\alias{IndexW}
\title{W index value for a single topology.}
\usage{
IndexW(tree = tree)
}
\arguments{
\item{tree}{is a single tree with n terminals, an ape phylo object.}
}
\description{
This function assigns the weights according to the ramification pattern [1]. 
The input tree is reordered in post order.
Returns a vector with weights.
}
\examples{
  library(jrich)
  data(tree)
  plot(tree)
  indexw             <- IndexW(tree)
  newTree            <- tree
  newTree$tip.label  <- indexw
  plot(newTree)
  


}
\author{
Miranda-Esquivel Daniel R.

#'@references
[1] Van-Wright et al., 1991. What to Protect? Systematics and the Agony of Choice.
Biological Conservation 55 (1991) 235-254
}

