% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Jack.Index}
\alias{Jack.Index}
\title{Jack-knife indices in a single topology}
\usage{
Jack.Index(tree = tree, distrib = distrib, jtip = 0, verbose = TRUE)
}
\description{
The function jack-knifes the terminals and calculates the indices.
}

