% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Jack.Indices}
\alias{Jack.Indices}
\title{Jack-knife indices in a single topology n times}
\usage{
Jack.Indices(tree = tree, distrib = distrib, jtip = jtip,
  replicates = replicates)
}
\description{
The function jack-knifes the terminals and
calculates the indices value n (=replicates) times.
The function returns the success that correspond to
obtain the same ranking for 1:3 positions.
}

