% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MultiJack}
\alias{MultiJack}
\title{Jack-knife indices in n topologies  m times}
\usage{
MultiJack(MultiData = MultiData, times = 100, jtip = 0, jtopol = 0)
}
\description{
The function calculates the indices values for a MultiData list  n (=replicates) times
}

