% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multitaxon1.R
\docType{data}
\name{Multitaxon1}
\alias{Multitaxon1}
\title{Multitaxon example}
\format{A list with two objects, each with an APE object, class phylo and the distribution of the terminals.}
\usage{
Multitaxon1
}
\description{
A multitaxon example with two hypothetical taxa / distributions
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}

