% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum.indices.2.topologies.R
\name{Sum.Indices.2.Topologies}
\alias{Sum.Indices.2.Topologies}
\title{Sums two tables with indices values}
\usage{
Sum.Indices.2.Topologies(indices1 = indices1, indices2 = indices2)
}
\arguments{
\item{indices1}{a table of indices values.}

\item{indices2}{a table of indices values.}
}
\value{
a single table with the sum of the two indices.
}
\description{
You input two tables with indices 
and it function returns a single table with the sum.
}
\author{
Miranda-Esquivel Daniel R.
}

