% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{taxon1}
\alias{taxon1}
\title{Figure 1 tree and distribution}
\format{A list with two objects}
\usage{
taxon1
}
\description{
A list with two objects:
Taxon[[1]]: Figure 1 tree, an APE object with five terminals and
Taxon[[2]]: A data frame with five species distributions in eight areas.
}
\keyword{datasets}

