% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{tree.figure1}
\alias{tree.figure1}
\title{Figure 1 tree}
\format{\preformatted{List of 3
 $ edge     : int [1:8, 1:2] 6 6 7 7 8 8 9 9 1 7 ...
 $ tip.label: chr [1:5] "I" "II" "III" "IV" ...
 $ Nnode    : int 4
 - attr(*, "class")= chr "phylo"
 - attr(*, "order")= chr "cladewise"
 - attr(*, "comment")= chr [1:2] "An hypothetical tree" "From: Miranda-Esquivel 2015"
}}
\usage{
tree.figure1
}
\description{
An APE object, class phylo. A tree dataset with five terminals.
}
\details{
A phylo object. A tree with five terminals.
}
\keyword{datasets}

