% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{Figure 1 tree}
\format{An object of class \code{phylo} of length 3.}
\usage{
tree
}
\description{
An APE object, class phylo. A tree dataset with five terminals.
}
\details{
A phylo object. A tree with five terminals.
}
\keyword{datasets}

