% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wconsensus.R
\name{wconsensus}
\alias{wconsensus}
\title{Wconsensus}
\usage{
wconsensus(InitialTrees, collapse = TRUE, cutvalue = 0.5001)
}
\description{
wconsensus calculates the weighted consensus as presented in Sharkey et al. (2013)
The initial trees are hierarchically and nested clustered, according to common nodes
and weighted in accordance to the clusters created.

The function receives as input data the initial trees (as a multiphylo object) 
and returns a collapsed tree according to the calculated weights, the default cut
value is 0.5

Usage:

ConsensusTree <- wconsensus (trees)
calculates  the weights and collapses the maj rule consensus
ConsensusTree <- wconsensus (trees, collapse = FALSE)
calculates the weights BUT does not collapse the maj rule consensus
ConsensusTree <- wconsensus (trees, cutvalue = 0.65)
calculates  the weights and collapses the maj rule consensus using
a cut value of 0.65
}
\author{
Miranda-Esquivel Daniel R.
}

