% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Weight.other.nodes}
\alias{Weight.other.nodes}
\title{Weight the tips that are sisters of internal nodes}
\usage{
Weight.other.nodes(tree = tree, matriz = matriz)
}
\description{
Input: a tree and a table of weights
Returns a distribution
}

