% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Best.Index}
\alias{Best.Index}
\title{Jack-knife indices in a single topology m times and evaluates a
success rule.}
\usage{
Best.Index(tree = tree, distrib = distrib, jtip = jtip,
  replicates = replicates, success = c(success))

Best.Index(tree = tree, distrib = distrib, jtip = jtip,
  replicates = replicates, success = c(success))
}
\description{
The function jack-knifes the terminals and
calculates the indices value m (=replicates) times.
The function returns the success that correspond to
obtain the same ranking for X,Y positions, established as the vector
success (by default success is 1:3 positions)).

The function jack-knifes the terminals and
calculates the indices value m (=replicates) times.
The function returns the success that correspond to
obtain the same ranking for X,Y positions, established as the vector
success (by default success)).
}

