% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.Index.R
\name{Calculate.Index}
\alias{Calculate.Index}
\title{Indices values and Jack-knife indices for a single topology.}
\usage{
Calculate.Index(tree = tree, distrib = distrib, jtip = 0,
  verbose = TRUE, standard = "distribution")
}
\arguments{
\item{tree}{is a single tree with n terminals, an ape phylo object.}

\item{distrib}{species distributions in n areas, a data.frame}

\item{jtip}{is the number of terminals, an integer.}

\item{verbose}{}

\item{standard}{}
}
\description{
The funtion calculates standard and terminal jack-knifed indices I and W 
[see Miranda-Esquivel 2015], along with Posadas et al. 2001 modifications.
}
\examples{
library(jrich)
data(tree)
data(distribution)
##
Standarized by the sum of indices in the distribution
Calculate.Index(tree=tree, distrib = distribution, verbose=TRUE, standard = "distribution")

## Standarized by the sum of indices in the tree (as figure 1 in Miranda-Esquivel 2015)
Calculate.Index(tree=tree, distrib = distribution, verbose=TRUE, standard = "tree")


Indices values and Jack-knife indices for a single topology.

The funtion calculates standard and terminal jack-knifed indices I and W 
[see Miranda-Esquivel 2015], along with Posadas et al. 2001 modifications.

}
\author{
Miranda-Esquivel Daniel R.
}

