% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Children.R
\name{Children}
\alias{Children}
\title{Children of a node.}
\usage{
Children(tree = tree, node = node)
}
\arguments{
\item{tree}{is a single tree}

\item{node}{is an integer, representing the node in APE notation}
}
\value{
The children nodes of the internal node; in most cases, two integers.
}
\description{
Get the children of a node in a tree.
}

