% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Children.R
\name{Children}
\alias{Children}
\title{Children of a node.}
\usage{
Children(tree = tree, node = node)
}
\arguments{
\item{tree}{is a single tree with n terminals, an ape phylo object.}

\item{node}{representing the node in APE notation,is an integer.}
}
\value{
The children nodes of the internal node; in most cases, two integers.
}
\description{
Get the children of a node in a tree.
}
\author{
Miranda-Esquivel Daniel R.
}

