% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\docType{data}
\name{distribution}
\alias{distribution}
\title{Figure 1 distribution}
\format{A data.frame object with eight areas (columns) and five species (rows)}
\usage{
distribution
}
\description{
A data frame with five species distributions in eight areas.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}

