% Generated by roxygen2 (4.0.1): do not edit by hand
\name{IndexI}
\alias{IndexI}
\title{I index value for a single topology.

This function assigns the same weight to sister clades
(see Van-Wright et al., 1981).
The input tree is reordered in post order.

Returns a vector with weights.}
\usage{
IndexI(tree = tree)
}
\description{

}
\examples{
library(jrich)
 data(tree)
 plot(tree)
 indexi               <- IndexI(tree)
 newTree              <- tree
 newTree$tip.label    <- indexi
 plot(newTree)
}

