% Generated by roxygen2 (4.0.1): do not edit by hand
\name{IndexI}
\alias{IndexI}
\title{I index value for a single topology}
\usage{
IndexI(tree = tree)
}
\description{
This function assigns the same weight to sister clades
(see Van-Wright et al., 1981).
The input tree is reordered in post order.
}
\details{
Returns a vector with weights.
}
\examples{
#indexi <- IndexI(tree)
#newTree <- tree
#newTree$Labels  <- indexi
#plot(newTree)
}

