% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IndexW}
\alias{IndexW}
\title{W index value for a single topology}
\usage{
IndexW(tree = tree)
}
\description{
This function assigns the weight according to the ramification patters
(see Van-Wright et al., 1981).
The input tree is reordered in post order.
}
\details{
Returns a vector with weights.
}
\examples{
# library(jrich)
 # data(tree.figure1)
 # indexw <- IndexW(tree.figure1)
 # newTree <- tree.figure1
 # newTree$Labels  <- indexw
 # plot(newTree)
}

