% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Jack.Index}
\alias{Jack.Index}
\title{Indices values and Jack-knife indices for a single topology}
\usage{
Jack.Index(tree = tree, distrib = distrib, jtip = 0, verbose = TRUE,
  standard = "distribution")
}
\description{
The funtion calculates standard and terminal jack-knifed indices I and W,
along with Posadas et al. (2001) modifications.
}

