% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Multi.Index.Calc}
\alias{Multi.Index.Calc}
\title{Jack-knife indices in n topologies one time}
\usage{
Multi.Index.Calc(MultiData = MultiData, jtip = 0, jtopol = 0)
}
\description{
The function calculates the indices values for a MultiData list
one time
}

