% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi.Jack.R
\name{Multi.Jack}
\alias{Multi.Jack}
\title{Jack-knife indices in n topologies  m times}
\usage{
Multi.Jack(MultiData = MultiData, times = 100, jtip = 0, jtopol = 0)
}
\description{
The function calculates the indices values for a MultiData list  m (=replicates) times 

Returns the rankings
}

