% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multitaxon1.R
\docType{data}
\name{Multitaxon1}
\alias{Multitaxon1}
\title{Multitaxon list}
\format{A list with two objects}
\usage{
Multitaxon1
}
\description{
A list with two objects:
MultiTaxon1[[1]]: A tree and its distribution for taxon 1
MultiTaxon1[[2]]: A tree and its distribution for taxon 2
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}

