% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Rank.Indices}
\alias{Rank.Indices}
\title{Rank indices according to the areas' absolute position.
If the index value is empty, the function assigns a dummy position "X0X"}
\usage{
Rank.Indices(index.Value = index.Value)
}
\description{
Input a data matrix with the index values
Return a table with the dreacins order of the areas
It presents the ties alphabetically
}

