% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.data.R
\name{Read.Data}
\alias{Read.Data}
\title{Read distributions as a csv with two columns species and area}
\usage{
Read.Data(data.File)
}
\description{
Read distributions as a csv with two columns species and area
}

