% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.data.R
\name{Read.Data}
\alias{Read.Data}
\title{Read distributions}
\usage{
Read.Data(data.File)
}
\arguments{
\item{data.File}{a csv file to read}
}
\value{
a data.frame object with the distribution by species
}
\description{
Read distributions as a csv with two columns species and area
}
\author{
Miranda-Esquivel Daniel R.
}

