% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum.indices.2.topologies.R
\name{Sum.Indices.2.Topologies}
\alias{Sum.Indices.2.Topologies}
\title{Sums two tables with indices values}
\usage{
Sum.Indices.2.Topologies(indices1 = indices1, indices2 = indices2)
}
\description{
Input: two tables with indices
Returns a single table with the sum
}

