% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon1.R
\docType{data}
\name{taxon1}
\alias{taxon1}
\title{A tree and the distribution of the taxa}
\format{A list with two objects}
\usage{
taxon1
}
\description{
A list with two objects:
Taxon[[1]]: A tree as an APE object with five terminals and
Taxon[[2]]: A data frame with five species distributions in eight areas,
following Miranda-Esquivel (2016).
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}

