% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{Figure 1 tree}
\format{An APE object, class phylo. A tree dataset with five terminals.}
\usage{
tree
}
\description{
A tree with five terminals.
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}

