% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{A tree with five terminals}
\format{An APE object, class phylo. A tree dataset with five terminals.}
\usage{
tree
}
\description{
A tree as an APE object with five terminals,
following Miranda-Esquivel (2016).
}
\author{
Miranda-Esquivel Daniel R.
}
\keyword{datasets}

